requirements.txt
streamlit
openai
chromadb
python-dotenv